global: split library in specific purpose folders
