+++ src/lib/elements/index.js
@@ -1,9 +1,10 @@
 // This file is part of React-Invenio-Forms
-// Copyright (C) 2020 CERN.
-// Copyright (C) 2020 Northwestern University.
+// Copyright (C) 2022 CERN.
 //
 // React-Invenio-Forms is free software; you can redistribute it and/or modify it
 // under the terms of the MIT License; see LICENSE file for more details.
 
-export * from './elements';
-export * from './forms';
+/**
+ * This folder contains general purpose reusable components.
+ */
+export { Image } from './Image';
