+++ src/lib/forms/index.js
@@ -1,0 +1,27 @@
+// This file is part of React-Invenio-Forms
+// Copyright (C) 2020-2022 CERN.
+// Copyright (C) 2020-2022 Northwestern University.
+//
+// React-Invenio-Forms is free software; you can redistribute it and/or modify it
+// under the terms of the MIT License; see LICENSE file for more details.
+
+/**
+ * This folder contains reusable components that can be used a form.
+ * The components are assuming that formik is used as a form creation library.
+ */
+export { AccordionField } from './AccordionField';
+export { ArrayField } from './ArrayField';
+export { BaseForm } from './BaseForm';
+export { BooleanField } from './BooleanField';
+export { ErrorLabel } from './ErrorLabel';
+export { ErrorMessage } from './ErrorMessage';
+export { FieldLabel } from './FieldLabel';
+export { GroupField } from './GroupField';
+export { SelectField } from './SelectField';
+export { RemoteSelectField } from './RemoteSelectField';
+export { TextField } from './TextField';
+export { TextAreaField } from './TextAreaField';
+export { ActionButton } from './ActionButton';
+export { RadioField } from './RadioField';
+export { RichInputField } from './RichInputField';
+export { ToggleField } from './ToggleField';
